Refactor playwright screenshot files
